// Çağrı Merkezi Angular/src/app/entities/models/cagri.model.ts
export class Cagri { 
    cagri_id: number = 0; 
    musteri_ad?: string; 
    musteri_soyad?: string; 
    musteri_telNo?: string; 
    musteri_email?: string; 
    sorun_basligi?: string; 
    sorun_aciklamasi?: string; 
    kayit_tarihi?: Date; 
    termin_tarihi?: Date; 
    servis_tarihi?: Date; 
    servis_ucreti?: number; 
    personel_id: number = 0; 
    urun_grubu: number = 0;
    cihaz_ismi?: string; 
} 